fix(db): money columns int -> bigint, and the narrowing deleted rather than made loud
What: all thirteen money columns across otc_orders and otc_billing widened to
bigint, every (int) narrowing cast on a money value removed, and a schema guard
added so the regression cannot return silently.

This corrects an error of the leader's carried through three phases. The plan
justified int columns as "spec parity" with #7. That justification was wrong:
specs/shared/requirements.md R1 says only "an integer count of minor units" and
domain-model.md M1 says "integer minor units only" — neither specifies a width,
so bigint satisfies both. int was #7's MySQL choice, recorded in a reference
document that describes #7's implementation rather than the shared contract,
and #7 needs no narrowing at all because JavaScript numbers are doubles.

Keeping int forced a long -> int narrowing at every money boundary that only #8
would pay for. The spec author proposed making that narrowing checked, which
was right given the premise; the human gate removed the premise instead. Three
of those casts were already in src/Seed and were unchecked — silently
truncating money today.

The Phase 6 migrations were amended in place rather than superseded, at the
human gate: they never left this machine, nothing depends on them, every test
run recreates from empty, and a permanent ALTER scar documenting a three-day-old
mistake is worse than a clean schema plus an honest record.

Verified: the seed's 413-row master-data parity against #7's live database is
unchanged, same SHA-256 as the Phase 7 baseline, so widening altered no value.

This is the second time this build mistook a detail of #7's implementation for
a requirement of the shared specification. The first was JSON payload key
ordering in Phase 5. Recorded as such in progress/history.md, because the
pattern is worth more to #9 than the fix.

Packages installed: none.

+++ src/Billing/Infrastructure/Persistence/Entities/Credit.cs
@@ -17,7 +17,7 @@ public sealed class Credit
 
     public string CompanyCode { get; set; } = string.Empty;
 
-    public int CreditLimit { get; set; }
+    public long CreditLimit { get; set; }
 
     public string CurrencyCode { get; set; } = string.Empty;
